refactor: move edgeconfig hook into common hooks dir
