+++ src/common/hooks/useEdgeConfig.ts
@@ -1,0 +1,4 @@
+import { EdgeConfigContext } from "@/common/providers/EdgeConfig";
+import { useContext } from "react";
+
+export const useEdgeConfigs = () => useContext(EdgeConfigContext);
