+++ src/common/providers/EdgeConfig/index.ts
@@ -1,1 +1,2 @@
 export * from "./EdgeConfigProvider";
+export * from "./EdgeConfigContextProvider";
